CLN: 컴포넌트명 정리
